reorganize tests to use helper file for basic and full prvate transfer
